Reformatted dtd-metrics code
