+++ jobs/dtd-metrics/src/metrics/__tests__/tenant-signups-trend.metric.test.ts
@@ -2,8 +2,8 @@ import { getTenantMock } from '@interop-be-reports/commons'
 import { randomUUID } from 'crypto'
 import { MacroCategoryCodeFor, MacroCategoryName, readModelMock, seedCollection } from '../../utils/tests.utils.js'
 import { sub } from 'date-fns'
-import { getTenantSignupsTrendMetric } from '../tenant-signups-trend-metric.service.js'
-import { GlobalStoreService } from '../global-store.service.js'
+import { GlobalStoreService } from '../../services/global-store.service.js'
+import { tenantSignupsTrendMetric } from '../tenant-signups-trend.metric.js'
 
 const comuneAttributeUuid = randomUUID()
 const aziendaOspedalieraAttributeUuid = randomUUID()
@@ -11,7 +11,7 @@ const aziendaOspedalieraAttributeUuid = randomUUID()
 const oneMonthAgoDate = sub(new Date(), { months: 1 }).toISOString()
 const sixMonthsAgoDate = sub(new Date(), { months: 6 }).toISOString()
 const oneYearAgoDate = sub(new Date(), { years: 1 }).toISOString()
-describe('getTenantSignupsTrendMetric', () => {
+describe('tenantSignupsTrendMetric', () => {
   it('should return the correct metrics', async () => {
     const oboardedTenants = [
       {
@@ -66,13 +66,13 @@ describe('getTenantSignupsTrendMetric', () => {
     await seedCollection('attributes', attributes)
 
     const globalStore = await GlobalStoreService.init(readModelMock)
-    const result = await getTenantSignupsTrendMetric(globalStore)
+    const result = await tenantSignupsTrendMetric.factoryFn(readModelMock, globalStore)
 
-    const comuniMetric = result.fromTheBeginning.find(
+    const comuniMetric = result?.fromTheBeginning.find(
       (metric) => metric.name === ('Comuni' satisfies MacroCategoryName)
     )
 
-    const aziendeOspedaliereMetric = result.fromTheBeginning.find(
+    const aziendeOspedaliereMetric = result?.fromTheBeginning.find(
       (metric) => metric.name === ('Aziende Ospedaliere e ASL' satisfies MacroCategoryName)
     )
