+++ jobs/dtd-metrics/src/metrics/published-e-services.metric.ts
@@ -1,12 +1,12 @@
-import { EServiceDescriptor, ReadModelClient } from '@interop-be-reports/commons'
+import { EServiceDescriptor } from '@interop-be-reports/commons'
 import { PublishedEServicesMetric } from '../models/metrics.model.js'
-import { getMonthsAgoDate, getVariationPercentage } from '../utils/helpers.utils.js'
+import { getMonthsAgoDate, getVariationPercentage, wrapMetricFactoryFn } from '../utils/helpers.utils.js'
 import { Document } from 'mongodb'
 
 /**
  * @see https://pagopa.atlassian.net/browse/PIN-3744
  **/
-export async function getPublishedEServicesMetric(readModel: ReadModelClient): Promise<PublishedEServicesMetric> {
+export const publishedEServicesMetric = wrapMetricFactoryFn('publishedEServices', async (readModel) => {
   const oneMonthAgoDate = getMonthsAgoDate(1)
   const twoMonthsAgoDate = getMonthsAgoDate(2)
 
@@ -52,4 +52,4 @@ export async function getPublishedEServicesMetric(readModel: ReadModelClient): P
   const variation = getVariationPercentage(lastMonthCount, twoMonthsAgoPublishedEServicesCount)
 
   return PublishedEServicesMetric.parse({ count, lastMonthCount, variation })
-}
+})
