+++ jobs/dtd-metrics/src/services/index.ts
@@ -1,10 +1,2 @@
-export * from './e-services-by-macro-categories-metric.service.js'
-export * from './published-e-services-metric.service.js'
-export * from './most-subscribed-e-services-metric.service.js'
-export * from './top-producers-by-subscribers-metric.service.js'
-export * from './onboarded-tenants-count.service.js'
-export * from './tenant-distribution-metric.service.js'
-export * from './tenant-signups-trend-metric.service.js'
-export * from './onboarded-tenants-count-by-macro-categories-metric.service.js'
-export * from './top-producers-metric.service.js'
+export * from './github-client.service.js'
 export * from './global-store.service.js'
